Ajoute un service d’inscription des utilisateurs
